Add files via upload
This project creates a program that creates a cutting-edge virtual try-on platform for fashion, revolutionizes the online shopping experience. Using advanced augmented reality and computer vision techniques, it allows users to virtually try on clothing and accessories in real-time, providing a seamless and interactive way to visualize how different items fit and look before making a purchase, enhancing confidence and satisfaction in online fashion choices.